requirements.txt
torch>=2.2.0
numpy>=1.21.0
optuna>=3.6.0
pandas>=1.3.0
matplotlib
scikit-learn 
seaborn 
pyserial>=3.5
PyQt5>=5.15.0
matplotlib>=3.7.0 